move tailscale code to a seperate file add a method to create tailscale/headscale client by type define service methods for join vpn and status
