+++ packages/syft/src/syft/service/network/network_service.py
@@ -62,7 +62,6 @@ class HTTPNodeRoute(SyftObject, NodeRoute):
     private: bool = False
     protocol: str = "http"
     port: int = 80
-    vpn_key: str
 
     def __hash__(self) -> int:
         return (
