+++ packages/syft/src/syft/service/vpn/vpn_service.py
@@ -1,10 +1,16 @@
 # stdlib
+import os
 from typing import Optional
+from typing import Type
 
 # third party
+from result import Err
+from result import Ok
+from result import Result
 from sympy import Union
 
 # relative
+from ...client.client import SyftClient
 from ...serde.serializable import serializable
 from ...store.document_store import DocumentStore
 from ...types.syft_object import SYFT_OBJECT_VERSION_1
@@ -17,6 +23,13 @@ from ..response import SyftError
 from ..response import SyftSuccess
 from ..service import AbstractService
 from ..service import service_method
+from ..user.user_roles import GUEST_ROLE_LEVEL
+from .headscale_client import BaseVPNClient
+from .headscale_client import HeadScaleAuthToken
+from .headscale_client import HeadScaleClient
+from .headscale_client import VPNClientConnection
+from .tailscale_client import TailScaleClient
+from .tailscale_client import TailscaleStatus
 from .vpn_stash import VPNStash
 
 
@@ -27,6 +40,28 @@ class NodePeerVPNKey(SyftObject):
     id: Optional[UID]
     auth_key: str
     name: str
+
+
+def get_vpn_client(
+    client_type: Type[BaseVPNClient],
+) -> Result[BaseVPNClient, str]:
+    api_key = os.getenv("STACK_API_KEY")
+
+    url = None
+
+    if isinstance(client_type, HeadScaleClient):
+        url = os.getenv("TAILSCALE_URL", "http://proxy:4000")
+    elif isinstance(client_type, TailScaleClient):
+        url = os.getenv("HEADSCALE_URL", "http://headscale:4000")
+
+    if api_key and url:
+        client = client_type(
+            connection=VPNClientConnection(url=url),
+            api_key=api_key,
+        )
+        return Ok(client)
+
+    return Err(f"Failed to create client for: {client_type.__name__}")
 
 
 @instrument
@@ -43,28 +78,60 @@ class VPNService(AbstractService):
     def join_vpn(
         self,
         context: AuthedServiceContext,
+        client: SyftClient,
     ) -> Union[SyftSuccess, SyftError]:
         """Join a VPN Service"""
-        # node_id = node.node.create_or_get_node(  # type: ignore
-        #     node_uid=res_json["node_id"],
-        #     node_name=res_json["node_name"],
-        # )
 
-        # self.stash.add_vpn_endpoint(  # type: ignore
-        #     context=context
-        #     host_or_ip=res_json["host_or_ip"],
-        #     vpn_endpoint=str(grid_url.with_path("/vpn")),
-        #     vpn_key=res_json["vpn_auth_key"],
-        # )
-        pass
+        auth_token = client.vpn.register()
+
+        if isinstance(auth_token, SyftError):
+            return auth_token
+
+        result = get_vpn_client(TailScaleClient)
+
+        if result.is_err():
+            return SyftError(message=result.err())
+
+        tailscale_client = result.ok()
+
+        result = tailscale_client.disconnect()
+
+        if isinstance(result, SyftError):
+            return result
+
+        result = tailscale_client.connect(
+            headscale_host=client.connection.url,
+            headscale_auth_token=auth_token.key,
+        )
+
+        if isinstance(result, SyftError):
+            return result
+
+        result = self.stash.add_vpn_endpoint(
+            context=context,
+            host_or_ip=client.connection.url,
+            vpn_key=auth_token.key,
+        )
+
+        if result.is_err():
+            return SyftError(message=result.err())
+
+        return SyftSuccess(message="Successfully joined VPN !!!")
 
     @service_method(path="vpn.status", name="status")
     def get_status(
         self,
         context: AuthedServiceContext,
-    ) -> Union[SyftSuccess, SyftError]:
+    ) -> Union[TailscaleStatus, SyftError]:
         """Join a VPN Service"""
-        pass
+        result = get_vpn_client(TailScaleClient)
+
+        if result.is_err():
+            return SyftError(message=result.err())
+
+        tailscale_client = result.ok()
+
+        return tailscale_client.status()
 
     def connect_vpn(
         self,
@@ -73,9 +140,20 @@ class VPNService(AbstractService):
         """Join a VPN Service"""
         pass
 
+    @service_method("register", "vpn.register", roles=[GUEST_ROLE_LEVEL])
     def register(
         self,
         context: AuthedServiceContext,
-    ) -> Union[SyftSuccess, SyftError]:
+    ) -> Union[HeadScaleAuthToken, SyftError]:
         """Register node to the VPN."""
-        pass
+
+        result = get_vpn_client(HeadScaleClient)
+
+        if result.is_err():
+            return SyftError(message=result.err())
+
+        headscale_client = result.ok()
+
+        token = headscale_client.generate_token()
+
+        return token
